fixed connecting to a remote database based on environment variable.
If no environment variable is defined, use the default environment variable for surrealdb 127.0.0.1:8000.

If SURREAL_BIND is defined as environmentvariable, use this. It also soft checks if it just an ip address by checking if the first character is a digit. If this is the case, it adds ws:// infront of the ip address to connect over ws